Redo file structure
