+++ src/lib.rs
@@ -3,10 +3,6 @@
 #![doc(html_logo_url = "https://raw.githubusercontent.com/Hoverbear/raft/master/raft.png")]
 #![doc(html_root_url = "https://hoverbear.github.io/raft/raft/")]
 
-#![cfg_attr(feature = "serde", feature(plugin))]
-#![cfg_attr(feature = "serde", feature(custom_derive))]
-#![cfg_attr(feature = "serde", plugin(serde_macros))]
-
 //! This is the Raft Distributed Consensus Protocol implemented for Rust.
 //! [Raft](http://raftconsensus.github.io/) is described as:
 //!
